refactor(console): merge auth and shell store
