+++ apps/console/app/components/ui/UserName.tsx
@@ -1,9 +1,9 @@
 import { useSnapshot } from "valtio";
 
-import { authStore } from "~console/state/auth";
+import { shellStore } from "~console/state/auth";
 
 export function UserName() {
-  const { user } = useSnapshot(authStore);
+  const { user } = useSnapshot(shellStore);
 
   return <span>{user?.name}</span>;
 }
