--- apps/console/app/state/shell.ts
@@ -1,15 +1,0 @@
-import { proxy } from "valtio";
-
-export type Models = Record<
-  string,
-  {
-    name: string;
-    modality: string;
-    providers: string[];
-    monthlyFreeTokens: number;
-  }
->;
-
-export const shellStore = proxy<{ models: Models | undefined }>({
-  models: undefined,
-});
